Added "DistanceOutside" as a data category for the DNN.

+++ SVDModel/SVDCore/tools/expressionwrapper.cpp
@@ -42,7 +42,7 @@ void set_strings()
         return;
     // ********* base **********
     baseVarList.push_back("year");
-    baseVarListCount = baseVarList.size();
+    baseVarListCount = static_cast<int>(baseVarList.size());
 
     // ********* tree **********
